[refactor] Move GUI from misc to ui.
First move GUI class into taichi.ui module. We could possibly structure
this better but it won't matter too much since GUI is exposed at `ti`
level.

[ghstack-poisoned]
