+++ python/taichi/ui/__init__.py
@@ -1,1 +1,2 @@
+from .gui import *
 from .ui import *
